PI-1415 add appointment retrieval endpoint
