+++ projects/resettlement-passport-and-delius/build.gradle.kts
@@ -33,6 +33,7 @@ dependencies {
 configure<ClassPathExtension> {
     jacocoExclusions = listOf(
         "**/config/**",
+        "**/model/**",
         "**/entity/**",
         "**/AppKt.class"
     )
